added handshake

+++ events/messageCreate.js
@@ -2,7 +2,8 @@
     This event is triggers by Discord and does processing of data  */
 
 //require packages
-const { getGuildPrefix, getCommandPermissions, checkCommandPermissions } = require("../utils/PermissionManager");
+const { HandshakeMessage } = require("../utils/MessageManager");
+const { getCommandPermissions, checkCommandPermissions } = require("../utils/PermissionManager");
 
 //exports "message" event
 module.exports = async (client, message) => {
@@ -23,6 +24,14 @@ module.exports = async (client, message) => {
 
     //check if content starts with prefix, else return
     if (messagePrefix.startsWith(prefix)) {
+
+        //check if bot has been activated, else return
+        if (message.guild.handshake == null) {
+            //if member is moderator, return handshake message
+            if (message.member.permissions.has("MANAGE_GUILD")) return await HandshakeMessage(message, 4800);
+            else return;
+        }
+
         //check for regular command (including alliasses)
         const commandFile = (client.commands.get(messageCommand)) ?
             client.commands.get(messageCommand) :
@@ -38,13 +47,20 @@ module.exports = async (client, message) => {
             if (verification.status === true) {
                 await message.react('701401045473165352'); //react to command
                 commandFile.run(client, message, messageArgs, prefix, verification); //execute command
-            } /* else message.reply(verification.message); */
+            } // else message.reply(verification.message);
         }
     }
 
     //if client is mentioned, but no content is given return info
     if (message.content.startsWith('<@') && message.content.endsWith('>')) {
         if (message.mentions.users.first().id === client.user.id && messageArgs.length < 1) {
+            //check if bot has been activated, else return
+            if (message.guild.handshake == null) {
+                //if member is moderator, return handshake message
+                if (message.member.permissions.has("MANAGE_GUILD")) return await HandshakeMessage(message, 4800);
+                else return;
+            }
+            //reply with server info
             message.reply(`Hello, your current server prefix is \`${prefix}\``)
                 .then(msg => { setTimeout(() => msg.delete(), 5000) })
         }
